+++ src/app/app.component.ts
@@ -1,22 +1,23 @@
 import {Component} from '@angular/core';
-import {HomeComponent} from './home/home.component';
+import {RouterModule} from '@angular/router';
 
 
 @Component({
   selector: 'app-root',
-  imports: [HomeComponent],
+  imports: [RouterModule],
   template: `
     <main>
       <header class="brand-name">
         <img class="brand-logo" src="/assets/logo.svg" alt="logo" aria-hidden="true" />
       </header>
       <section class="content">
-        <app-home></app-home>
+        <router-outlet></router-outlet>
       </section>
     </main>
   `,
   styleUrls: ['./app.component.css'],
 })
+
 export class AppComponent {
   title = 'homes';
 }